Fix bug for keyer_app on Safari
On Safari, you can only create an audio context in response
to a user action, like clicking a button. Reworked code to
support that. Icky use of globals; this probably isn't the
best practice, but it works.

Also, close web audio context on cancel button, so we don't
leak audio contexts.