Move message sending logic to radio
Make the radio responsible for sending messages. It's just
cleaner, since the radio can mute itself, and can also tell
the band what frequency it's tuned to when sending CQ.

Add a "vox delay" so that we don't unkey immediately after
finishing a message. That makes the simulator act more
realistic.